+++ apps/website/src/components/explore/details/DefinitionItem.tsx
